Breaking Change: Use name as default for role_name, policy_name and instance_profile_name
This adds role_name and role_name_profix variables.

This changes the behavior of the name variable to be used as the default
for role_name, policy_name and instance_profile_name.

This changes the behavior of the name_prefix variable to be used as the default
for role_name_prefix, policy_name_prefix and instance_profile_name_prefix.

when upgrading change name to role_name and name_prefix to role_name_prefix.